+++ src/Components/index.js
@@ -1,2 +1,3 @@
 export { default as Layout } from './Layout';
 export { default as Navbar } from './Navbar';
+export { default as MissionList } from './MissionList';
